requirements.txt
pyblock==0.6.0
numba==0.51.2
pandas==1.0.3
colorama==0.3.9
scipy==1.4.1
icc-rt==2020.0.133
matplotlib==3.2.1
pyyaml==5.3.1
psutil==5.7.2
